use common custom project typing across project
