+++ shared/entities/custom-project.entity.ts
@@ -10,7 +10,7 @@ import { ACTIVITY } from "@shared/entities/activity.enum";
 import { Country } from "@shared/entities/country.entity";
 import { User } from "@shared/entities/users/user.entity";
 import { type CustomProjectOutput } from "@shared/dtos/custom-projects/custom-project-output.dto";
-import {CreateCustomProjectDto} from "@api/modules/custom-projects/dto/create-custom-project.dto";
+import {CreateCustomProjectDto} from "@shared/dtos/custom-projects/create-custom-project.dto";
 
 /**
  * @note: This entity does not extend BaseEntity as it won't be used in the backoffice. However, it has to be added to the BO datasource due to its relation
